Separate go build and gofmt checks.
