+++ test/unit/plugins/pre_commit/checks/go_fmt_test.rb
@@ -1,8 +1,8 @@
 require 'minitest_helper'
-require 'plugins/pre_commit/checks/go'
+require 'plugins/pre_commit/checks/go_fmt'
 
-describe PreCommit::Checks::Go do
-  let(:check) {PreCommit::Checks::Go.new(nil, nil, [])}
+describe PreCommit::Checks::GoFmt do
+  let(:check) {PreCommit::Checks::GoFmt.new(nil, nil, [])}
 
   it "succeds if nothing changed" do
     check.call([]).must_equal nil
@@ -16,7 +16,4 @@ describe PreCommit::Checks::Go do
     check.call([fixture_file("bad_fmt.go")]).must_match(/bad_fmt.go/)
   end
 
-  it "fails for compiler errors" do
-    check.call([fixture_file("dont_compile.go")]).must_match(/imported and not used/)
-  end
 end unless `which go 2>/dev/null`.empty?
